+++ src/main/java/roomescape/RoomescapeApplication.java
@@ -1,11 +1,12 @@
 package roomescape;
 
+import jwt.JwtConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
-@ConfigurationPropertiesScan(basePackages = "roomescape")
+@Import(JwtConfig.class)
 public class RoomescapeApplication {
     public static void main(String[] args) {
         SpringApplication.run(RoomescapeApplication.class, args);
