+++ src/main/java/roomescape/auth/AdminAuthInterceptor.java
@@ -1,5 +1,6 @@
 package roomescape.auth;
 
+import jwt.JwtTokenProvider;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
