+++ src/main/java/roomescape/member/presentation/MemberController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.AuthenticatedMember;
-import roomescape.auth.JwtProperties;
-import roomescape.auth.JwtTokenProvider;
+import jwt.JwtProperties;
+import jwt.JwtTokenProvider;
 import roomescape.auth.LoginMember;
 import roomescape.common.CookieUtil;
 import roomescape.member.application.MemberService;
